rag-chatbot/backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
python-dotenv>=1.0
python-multipart>=0.0.9
openai>=1.30
pypdf>=4.2
python-docx>=1.1
pytest>=8.0
httpx>=0.27
